requirements.txt
PyMuPDF==1.24.14
pydantic==2.10.3
openai==1.58.1
instructor==1.8.1
python-dotenv==1.0.1
